rename cyten_float->float64, cyten_int->int64, cyten_complex->complex128

+++ include/check.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cyten/cyten.h>
+
 namespace cyten {
 
 /**
  * A function adding the two arguments. Just there as a test of the surrounding setup.
  */
-int add(int i, int j);
+int64 add(int64 i, int64 j);
 
 }
